use clang-tidy

+++ tools/makedata/src/DataMaker/DataCreatureMaker.cpp
@@ -247,7 +247,7 @@ void DataCreatureMaker::makeDataCreature(
 
     bool issettraintime = false;
     for (auto index : earr::Enum<data::CreatureTrainTime>()) {
-        issettraintime |= data_creature_evolCondition.getTrainTime(index);
+        issettraintime |= static_cast<int>(data_creature_evolCondition.getTrainTime(index));
     }
     if (!issettraintime) {
         data_creature_evolCondition.setTrainTimeEveryDay();
